Implement SAI Authorization Agent and related endpoints (#367)
