+++ app-framework/app/package.json
@@ -21,6 +21,7 @@
     "form-data-encoder": "1.9.0",
     "formdata-node": "4.4.1",
     "handlebars": "^4.7.7",
+    "http-link-header": "^1.1.1",
     "moleculer": "^0.14.19",
     "sparqljs": "^3.5.2",
     "speakingurl": "^14.0.1",
