+++ app-framework/react/src/components/BackgroundChecks.tsx
@@ -1,9 +1,10 @@
-import { useCallback, useEffect, useState, useLayoutEffect, FunctionComponent, ReactNode } from 'react';
-import urlJoin from 'url-join';
-import { useGetIdentity, useNotify } from 'react-admin';
+import React, { useCallback, useEffect, useState, useLayoutEffect, FunctionComponent, ReactNode } from 'react';
+import { useGetIdentity, useDataProvider, useTranslate, useLogout } from 'react-admin';
+import { Box, CircularProgress, Typography, Button } from '@mui/material';
+import ErrorIcon from '@mui/icons-material/Error';
 import { useNodeinfo } from '@semapps/activitypub-components';
-import { arrayOf } from '../utils';
-import type { AppStatus } from '../types';
+import { arrayOf, delay } from '../utils';
+import useGetAppStatus from '../hooks/useGetAppStatus';
 
 /**
  * Call the /.well-known/app-status endpoint to check the status of the app
@@ -14,9 +15,13 @@ import type { AppStatus } from '../types';
  */
 const BackgroundChecks: FunctionComponent<Props> = ({ clientId, listeningTo = [], children }) => {
   const { data: identity, isLoading: isIdentityLoading } = useGetIdentity();
-  const notify = useNotify();
+  const dataProvider = useDataProvider();
+  const translate = useTranslate();
+  const logout = useLogout();
   const [appStatusChecked, setAppStatusChecked] = useState<boolean>(false);
+  const [errorMessage, setErrorMessage] = useState<string | undefined>();
   const nodeinfo = useNodeinfo(identity?.id ? new URL(identity?.id as string).host : undefined);
+  const getAppStatus = useGetAppStatus();
 
   const isLoggedOut = !isIdentityLoading && !identity?.id;
 
@@ -25,53 +30,64 @@ const BackgroundChecks: FunctionComponent<Props> = ({ clientId, listeningTo = []
   const checkAppStatus = useCallback(async () => {
     // Only proceed if the tab is visible
     if (!document.hidden && identity?.id) {
-      const oidcIssuer = new URL(identity?.id as 'string').origin;
-      const endpointUrl = urlJoin(oidcIssuer, '.well-known/app-status');
-      const token = localStorage.getItem('token');
-
       try {
-        // Don't use dataProvider.fetch as it would go through the proxy
-        const response = await fetch(endpointUrl, {
-          headers: new Headers({ Authorization: `Bearer ${token}`, Accept: 'application/json' })
-        });
-        if (response.ok) {
-          const appStatus = (await response.json()) as AppStatus;
-          if (appStatus) {
-            if (!appStatus.onlineBackend) {
-              notify('apods.error.app_offline', { type: 'error' });
-              return;
-            }
+        let appStatus = await getAppStatus();
+        if (appStatus) {
+          if (!appStatus.onlineBackend) {
+            setErrorMessage(translate('apods.error.app_offline'));
+            return;
+          }
 
-            if (!appStatus.installed) {
-              notify('apods.error.app_not_installed', { type: 'error' });
-              return;
-            }
+          if (!appStatus.installed) {
+            setErrorMessage(translate('apods.error.app_not_installed'));
+            return;
+          }
 
-            if (appStatus.upgradeNeeded) {
-              const consentUrl = new URL(nodeinfo?.metadata?.consent_url as 'string');
-              consentUrl.searchParams.append('client_id', clientId);
-              consentUrl.searchParams.append('redirect', window.location.href);
-              window.location.href = consentUrl.toString();
-              return;
-            }
+          if (appStatus.upgradeNeeded) {
+            const { json: actor } = await dataProvider.fetch(identity.id);
+            const { json: authAgent } = await dataProvider.fetch(actor['interop:hasAuthorizationAgent']);
+            const redirectUrl = new URL(authAgent['interop:hasAuthorizationRedirectEndpoint']);
+            redirectUrl.searchParams.append('client_id', clientId);
+            window.location.href = redirectUrl.toString();
+            return;
+          }
 
-            if (listeningTo.length > 0) {
+          if (listeningTo.length > 0) {
+            let numAttempts = 0,
+              missingListener;
+
+            do {
+              missingListener = undefined;
+
               for (const uri of listeningTo) {
                 if (!arrayOf(appStatus.webhookChannels).some(c => c.topic === uri)) {
-                  notify('apods.error.app_not_listening', { messageArgs: { uri }, type: 'error' });
-                  return;
+                  missingListener = uri;
                 }
               }
-            }
 
-            setAppStatusChecked(true);
+              // If one or more listener were not found, wait 1s and refetch the app status endpoint
+              // This happens when the app was just registered, and the webhooks have not been created yet
+              if (missingListener) {
+                numAttempts++;
+                await delay(1000);
+                appStatus = await getAppStatus();
+              }
+            } while (missingListener && numAttempts < 10);
+
+            if (missingListener) {
+              setErrorMessage(translate('apods.error.app_not_listening', { uri: missingListener }));
+              return;
+            }
           }
+
+          setAppStatusChecked(true);
         }
       } catch (e) {
-        notify('apods.error.app_status_unavailable', { type: 'error' });
+        console.error(e);
+        setErrorMessage(translate('apods.error.app_status_unavailable'));
       }
     }
-  }, [identity, nodeinfo, setAppStatusChecked, document]);
+  }, [identity, nodeinfo, getAppStatus, setAppStatusChecked, document, dataProvider, setErrorMessage, translate]);
 
   useEffect(() => {
     if (identity?.id && nodeinfo) {
@@ -86,11 +102,38 @@ const BackgroundChecks: FunctionComponent<Props> = ({ clientId, listeningTo = []
     return () => document.removeEventListener('visibilitychange', checkAppStatus);
   }, [checkAppStatus]);
 
-  // TODO display error message instead of notifications
   if (isLoggedOut || appStatusChecked) {
     return children;
+  } else if (errorMessage) {
+    return (
+      <Box display="flex" flexDirection="column" alignItems="center" justifyContent="center" sx={{ minHeight: 400 }}>
+        <Box sx={{ backgroundColor: 'red', p: 2, textAlign: 'center' }}>
+          <ErrorIcon sx={{ width: 50, height: 50, color: 'white' }} />
+          <Typography color="white">{errorMessage}</Typography>
+          <Button
+            variant="contained"
+            color="error"
+            sx={{ mt: 2, mr: 1 }}
+            onClick={() => {
+              setErrorMessage(undefined);
+              checkAppStatus();
+            }}
+          >
+            {translate('ra.action.refresh')}
+          </Button>
+          <Button variant="contained" color="error" sx={{ mt: 2 }} onClick={() => logout()}>
+            {translate('ra.auth.logout')}
+          </Button>
+        </Box>
+      </Box>
+    );
   } else {
-    return null;
+    // TODO wait 3s before display loader
+    return (
+      <Box display="flex" flexDirection="column" alignItems="center" justifyContent="center" sx={{ minHeight: 400 }}>
+        <CircularProgress size={100} thickness={6} sx={{ mb: 5, color: 'white' }} />
+      </Box>
+    );
   }
 };
