+++ pod-provider/backend/services/solid/solid-oidc.js
@@ -64,19 +64,6 @@ module.exports = {
       }
     });
 
-    await this.broker.call('api.addRoute', {
-      route: {
-        name: 'oidc-consent-completed',
-        path: path.join(basePath, '/.oidc/consent-completed'),
-        bodyParsers: {
-          json: true
-        },
-        aliases: {
-          'POST /': 'solid-oidc.consentCompleted'
-        }
-      }
-    });
-
     // The OIDC provider route must be added after all other routes, otherwise it gets overwritten
     // TODO find out why ! Probably due to the fact it is only a middleware (adding dummy aliases doesn't help)
     await delay(10000);
@@ -146,28 +133,6 @@ module.exports = {
       await this.interactionFinished(interactionId, {
         login: { accountId: webId, amr: ['pwd'], remember: true }
       });
-    },
-    // See https://github.com/panva/node-oidc-provider/blob/main/docs/README.md#user-flows
-    async consentCompleted(ctx) {
-      const { interactionId } = ctx.params;
-      const interaction = await this.oidc.Interaction.find(interactionId);
-
-      if (interaction) {
-        const clientId = interaction.params?.client_id;
-        const accountId = interaction.session?.accountId;
-        const scope = interaction.params?.scope;
-
-        // Grant the requested scope (should be "webid openid")
-        const grant = new this.oidc.Grant({ accountId, clientId });
-        grant.addOIDCScope(scope);
-        const grantId = await grant.save();
-
-        await this.interactionFinished(interactionId, {
-          consent: { grantId }
-        });
-      } else {
-        throw new Error(`No interaction found with ID ${interactionId}. It may have expired.`);
-      }
     }
   },
   methods: {
