+++ tests/initialize.js
@@ -10,7 +10,6 @@ const { TripleStoreAdapter } = require('@semapps/triplestore');
 const { WebAclMiddleware } = require('@semapps/webacl');
 const { interop, oidc, notify, apods } = require('@semapps/ontologies');
 const { NotificationsListenerService } = require('@semapps/solid');
-const { ACTIVITY_TYPES } = require('@semapps/activitypub');
 const RdfJSONSerializer = require('../pod-provider/backend/RdfJSONSerializer');
 const { clearMails } = require('./utils');
 const CONFIG = require('./config');
@@ -50,6 +49,25 @@ const clearDataset = dataset =>
     }
   });
 
+// Delete all data except special endpoints
+const clearSettingsDataset = () =>
+  fetch(`${CONFIG.SPARQL_ENDPOINT}settings/update`, {
+    method: 'POST',
+    body: `
+      DELETE {
+        ?subject ?predicate ?object .
+      } WHERE {
+        ?subject ?predicate ?object .
+        FILTER(STRSTARTS(STR(?subject), "urn:"))
+      }
+    `,
+    headers: {
+      'Content-Type': 'application/sparql-update',
+      'X-SemappsUser': 'system',
+      Authorization: `Basic ${Buffer.from(`${CONFIG.JENA_USER}:${CONFIG.JENA_PASSWORD}`).toString('base64')}`
+    }
+  });
+
 const clearRedisDb = async redisUrl => {
   const redisClient = new Redis(redisUrl);
   await redisClient.flushdb();
@@ -58,9 +76,11 @@ const clearRedisDb = async redisUrl => {
 
 const clearAllData = async () => {
   const datasets = await listDatasets();
-  for (let dataset of datasets) {
+  for (let dataset of datasets.filter(d => d != 'settings')) {
     await clearDataset(dataset);
   }
+
+  await clearSettingsDataset();
 
   await clearRedisDb(CONFIG.QUEUE_SERVICE_URL);
   await clearRedisDb(CONFIG.REDIS_OIDC_PROVIDER_URL);
@@ -177,7 +197,7 @@ const getAppAccessNeeds = async (actor, appUri) => {
   return [requiredAccessNeedGroup, optionalAccessNeedGroup];
 };
 
-const installApp = async (actor, appUri, waitForAccept = true, acceptedAccessNeeds, acceptedSpecialRights) => {
+const installApp = async (actor, appUri, acceptedAccessNeeds, acceptedSpecialRights) => {
   // If the accepted needs are not specified, use the app required access needs
   if (!acceptedAccessNeeds && !acceptedSpecialRights) {
     const [requiredAccessNeedGroup] = await getAppAccessNeeds(actor, appUri);
@@ -185,37 +205,11 @@ const installApp = async (actor, appUri, waitForAccept = true, acceptedAccessNee
     acceptedSpecialRights = requiredAccessNeedGroup['apods:hasSpecialRights'];
   }
 
-  const publishedAfter = new Date().toISOString();
-
-  // Do not await here
-  actor.call('activitypub.outbox.post', {
-    collectionUri: actor.outbox,
-    type: 'apods:Install',
-    object: appUri,
-    'apods:acceptedAccessNeeds': acceptedAccessNeeds,
-    'apods:acceptedSpecialRights': acceptedSpecialRights
+  await actor.call('auth-agent.install', {
+    appUri,
+    acceptedAccessNeeds,
+    acceptedSpecialRights
   });
-
-  const createRegistrationActivity = await actor.call('activitypub.outbox.awaitActivity', {
-    collectionUri: actor.outbox,
-    matcher: {
-      type: ACTIVITY_TYPES.CREATE,
-      to: appUri
-    },
-    publishedAfter
-  });
-
-  if (waitForAccept) {
-    await actor.call('activitypub.inbox.awaitActivity', {
-      collectionUri: actor.inbox,
-      matcher: {
-        type: ACTIVITY_TYPES.ACCEPT,
-        object: createRegistrationActivity.id
-      }
-    });
-  }
-
-  return createRegistrationActivity;
 };
 
 module.exports = {
